requirements.txt
alembic==1.7.1
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
pip==21.2.4
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.0
setuptools==56.0.0
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
